Dataset used for the analysis.
The Department of Education Statistics releases a data set annually containing the percentage of bachelor's degrees granted to women from 1970 to 2012. The data set is broken up into 17 categories of degrees, with each column as a separate category.
Randal Olson, a data scientist at the University of Pennsylvania, has cleaned the data set and made it available on his personal website.